refactor: move stylelint config and deps to tools/stylelint workspace
