+++ .stylelintrc.mjs
@@ -1,0 +1,5 @@
+/** @type {import('stylelint').Config} */
+export default {
+	extends: '@wordpress/stylelint-tools/config',
+	reportDescriptionlessDisables: true,
+};
