+++ package.json
@@ -67,6 +67,7 @@
 		"@typescript/native-preview": "^7.0.0-dev.20260423.1",
 		"@wordpress/eslint-tools": "file:./tools/eslint",
 		"@wordpress/release-tools": "file:./tools/release",
+		"@wordpress/stylelint-tools": "file:./tools/stylelint",
 		"@wordpress/vips": "file:./packages/vips",
 		"concurrently": "^3.5.0",
 		"cross-env": "^7.0.3",
@@ -77,7 +78,6 @@
 		"patch-package": "^8.0.0",
 		"prettier": "npm:wp-prettier@^3.0.3",
 		"rimraf": "^5.0.10",
-		"stylelint-plugin-logical-css": "^1.2.3",
 		"syncpack": "^15.3.1",
 		"typescript": "^6.0.2",
 		"wait-on": "^8.0.1"
