+++ packages/components/package.json
@@ -112,6 +112,7 @@
 		"@types/jest": "^29.5.14",
 		"@types/react-dom": "^18.3.1",
 		"@wordpress/jest-console": "file:../jest-console",
+		"@wordpress/stylelint-tools": "file:../../tools/stylelint",
 		"snapshot-diff": "^0.10.0",
 		"storybook": "^10.2.8",
 		"timezone-mock": "^1.3.6"
