+++ packages/theme/src/stylelint-plugins/test/utils/index.ts
@@ -4,11 +4,23 @@ import childProcess from 'node:child_process';
 
 const execute = util.promisify( childProcess.exec );
 
+/*
+ * Resolve Stylelint's binary through Node's module resolution rather than a
+ * fixed `node_modules/.bin` path, so the test works regardless of whether the
+ * dependency is hoisted to the repo root or isolated in this package.
+ */
+const stylelintBin = path.join(
+	path.dirname( require.resolve( 'stylelint/package.json' ) ),
+	'bin/stylelint.mjs'
+);
+
 const generateStylelintCommand = (
 	filename: string,
 	configFile: string
 ): string =>
-	'node_modules/.bin/stylelint ' +
+	'node ' +
+	stylelintBin +
+	' ' +
 	path.resolve( __dirname, '../', filename ) +
 	' -c ' +
 	path.resolve( __dirname, '../', configFile ) +
