+++ tools/stylelint/config.js
