+++ tools/stylelint/package.json
@@ -1,0 +1,31 @@
+{
+	"name": "@wordpress/stylelint-tools",
+	"version": "1.0.0",
+	"description": "Consolidated Stylelint configuration for the repo.",
+	"private": true,
+	"author": "The WordPress Contributors",
+	"license": "GPL-2.0-or-later",
+	"keywords": [
+		"wordpress",
+		"gutenberg",
+		"stylelint"
+	],
+	"homepage": "https://github.com/WordPress/gutenberg/tree/HEAD/tools/stylelint",
+	"repository": {
+		"type": "git",
+		"url": "git+https://github.com/WordPress/gutenberg.git",
+		"directory": "tools/stylelint"
+	},
+	"bugs": {
+		"url": "https://github.com/WordPress/gutenberg/issues"
+	},
+	"exports": {
+		"./config": "./config.js"
+	},
+	"dependencies": {
+		"stylelint-plugin-logical-css": "^1.2.3"
+	},
+	"publishConfig": {
+		"access": "public"
+	}
+}
